# docs/api/openapi.v1.json
{
  "components": {
    "headers": {
      "Retry-After": {
        "description": "How many seconds to wait before retrying.",
        "required": true,
        "schema": {
          "format": "int32",
          "type": "integer"
        }
      },
      "X-Correlation-Id": {
        "description": "The correlation identifier of this request. It is accepted on the request and echoed on every response, and it is the identifier a support call names.",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        },
        "description": "The request was not valid. The body is problem details with per-field `errors` when the failure is a validation failure.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/X-Correlation-Id"
          }
        }
      },
      "Forbidden": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        },
        "description": "The caller holds a session but not the permission, the branch reach or the session assurance this operation requires. The refusal is audited.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/X-Correlation-Id"
          }
        }
      },
      "InternalServerError": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        },
        "description": "The request failed unexpectedly. The body carries the problem type and the correlation identifier and nothing else — never a stack trace, never an exception message.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/X-Correlation-Id"
          }
        }
      },
      "NotFound": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        },
        "description": "There is no such resource, or none this caller may see. The two are deliberately indistinguishable, so that the identifier space cannot be probed.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/X-Correlation-Id"
          }
        }
      },
      "TooManyRequests": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        },
        "description": "The rate-limit policy this operation declares refused the request. `Retry-After` says when it is worth trying again.",
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/Retry-After"
          },
          "X-Correlation-Id": {
            "$ref": "#/components/headers/X-Correlation-Id"
          }
        }
      },
      "Unauthorized": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        },
        "description": "There is no live session, or the one presented has expired or been revoked. The client re-authenticates in place and retries the request with the same `Idempotency-Key`.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/X-Correlation-Id"
          }
        }
      },
      "UnsupportedMediaType": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        },
        "description": "The request body was not `application/json`.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/X-Correlation-Id"
          }
        }
      },
      "UpgradeRequired": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        },
        "description": "The client build declared in `X-Client-Version` is older than the minimum this server supports. Reloading collects the current build; repeating the request does not help. The body carries `minimumClient` and `current`.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/X-Correlation-Id"
          }
        }
      }
    },
    "schemas": {
...
      "ProblemDetails": {
        "description": "RFC 9457 problem details. Every failure is reported in this shape, and never as a stack trace, an exception message or a bare status code.",
        "properties": {
          "code": {
            "description": "The stable error code, `<module>.<kebab-case-reason>` — for example `idempotency.key-reused`. This is what a client branches on.",
            "type": "string"
          },
          "correlationId": {
            "description": "The correlation identifier of the request, echoed so a support call can name it.",
            "type": "string"
          },
          "currentEtag": {
            "description": "The same token in `ETag` form, ready to be sent back as `If-Match`.",
            "type": [
              "null",
              "string"
            ]
          },
          "currentVersion": {
            "description": "The aggregate's current concurrency token, on a version conflict, so the client can re-read and merge without a second round trip.",
            "type": [
              "null",
              "string"
            ]
          },
          "detail": {
            "description": "What went wrong on this occurrence, in words a person can act on.",
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "description": "Field errors, keyed by the field's name in the request payload. Present on validation failures only.",
            "type": [
              "null",
              "object"
            ]
          },
          "instance": {
            "description": "The request path this occurrence refers to.",
            "type": [
              "null",
              "string"
            ]
          },
          "retryable": {
            "description": "True when repeating the identical request may succeed. A retry of a command reuses the same `Idempotency-Key`.",
            "type": [
              "null",
              "boolean"
            ]
          },
          "status": {
            "description": "The HTTP status code, repeated so the body is self-contained.",
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type.",
            "type": "string"
          },
          "type": {
            "description": "The problem type, as `urn:tailor360:problem:<code>`. Stable across releases; a client may branch on it.",
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem details",
        "type": "object"
      },
      "ReasonPayload": {
        "properties": {
          "reason": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
...
      "StaffUserPayload": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "homeBranchId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "lastSignInAt": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "mfaEnrolment": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "displayName",
          "userName",
          "email",
          "status",
          "mfaEnrolment",
          "homeBranchId",
          "lastSignInAt",
          "createdAt",
          "version"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "antiForgeryToken": {
        "description": "The request half of the anti-forgery token pair, fetched from `GET /api/v1/antiforgery` and held in memory. Required on every state-changing request, including signing in: a sign-in forged from another origin would plant the attacker's account in the victim's browser. It is useless without the paired cookie, which script cannot read.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "sessionCookie": {
        "description": "An opaque server-side session identifier. The browser sends it automatically; script cannot read it, and no operation ever returns it. It is issued by signing in and is rotated on every change of assurance.",
        "in": "cookie",
        "name": "__Host-t360.session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The staff surface is `/api/v1/**`, served from the same origin as the client so that the\nsession travels in a cookie script cannot read. The major version lives in the URL and\nnothing else negotiates it; inside a major version only additive changes are permitted, and\na breaking one is refused by the diff gate unless it is approved and its deprecation is\nrecorded (`docs/architecture/conventions.md` section 5).\n\nEvery request and response carries `X-Correlation-Id`. Errors are RFC 9457 problem details\nwith a stable `code` and, for validation failures, per-field `errors`; no error ever carries\na stack trace or an exception message. Commands that a client may retry take an\n`Idempotency-Key`; editable aggregates carry an `ETag` and require `If-Match`.\n\nThis document is generated from the composed route table and committed to\n`docs/api/openapi.v1.json`. It is not written by hand, and an endpoint that does not appear\nin it fails the endpoint inventory test.",
    "summary": "The same-origin backend-for-frontend serving the HyFib Tailor 360 progressive web application.",
    "title": "HyFib Tailor 360 API",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
    "/api/v1/admin/users/{userId}": {
      "get": {
        "operationId": "GetStaffUser",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffUserPayload"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "426": {
            "$ref": "#/components/responses/UpgradeRequired"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Read one staff account for administration.",
        "tags": [
          "Administration"
        ],
        "x-tailor360-authorisation": "permission:admin.users",
        "x-tailor360-rate-limit": "default-user"
      }
    },
    "/api/v1/admin/users/{userId}/suspend": {
      "post": {
        "operationId": "SuspendStaffUser",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "reason": "Left the company on 5 September; access withdrawn at the manager's request."
              },
              "schema": {
                "$ref": "#/components/schemas/ReasonPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffUserPayload"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "426": {
            "$ref": "#/components/responses/UpgradeRequired"
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Precondition Required"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "antiForgeryToken": [],
            "sessionCookie": []
          }
        ],
        "summary": "Suspend a staff account and end its sessions.",
        "tags": [
          "Administration"
        ],
        "x-tailor360-audit": "identity.user.suspended",
        "x-tailor360-authorisation": "permission:admin.users",
        "x-tailor360-rate-limit": "write"
      }
    },
...
  },
  "servers": [
    {
      "description": "The origin serving the progressive web application.",
      "url": "/"
    }
  ],
...
}